+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
   as we want to treat it more like a cross-distribution fallback,
   rather than the primary source of packages (for now, at least)
 
+- env: make GOPATH and GOROOT variables optional
+
 ## [0.37.0] - 2020-06-09
 
 ### Removed